Split main/lib/test
Now we have separate files for the main driver build, the actual `main`
function which is very short, and the tests.
